% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeansimp.R
\name{Kmeansimp}
\alias{Kmeansimp}
\title{Improvement Kmeans}
\usage{
Kmeansimp(data, n, iter = 10)
}
\arguments{
\item{data}{The dataset needed to be clustered(type: matrix or dataframe)}

\item{n}{Number of clusters}

\item{iter}{Maximum iteration times for updating clustering result}
}
\value{
A list including all important information of clustering
}
\description{
This package is an improved version of Lloyd kmeans clustering
Standard sampling method is to randomly select k points( k cluster number)
as initial centroids. Yet this way usually converges too slow.
The improved version has 3 steps:
Step 1: From all objects randomly select one.
Step 2: select next initial centroids from n-objects in such a way that the
Euclidean distance of that object is maximum from other selected initial centroids.
Step 3: repeat step 2 until we get k initial centroids.
}
\examples{
library(Kmeansimp)
X = matrix(runif(100, 0, 100), 25, 4)
Kmeansimp(X, 3, iter = 15)
clu = Kmeansimp(X, 2)
clu$Clustering_Vector
}
